refactor(database-anonymisers): restructure database anonymisation and import/export functionality
- Removed the seed command and associated logic from the database command module.
- Consolidated database import and seeding logic into a dedicated importer module.
- Created separate modules for JSON and SQL dump management, including initialization and saving functionality.
- Introduced a new README file for the database anonymisers module, detailing usage and structure.
- Enhanced the anonymisation logic by separating concerns into dedicated functions for clarity and maintainability.
- Implemented utility functions for creating dump file paths and stringifying complex objects.
- Updated the database command module to handle new options for exporting data subsets and improved default file handling.
