+++ apps/backend-cli/src/commands/database.ts
@@ -28,6 +28,13 @@ export const databaseCommand: CommandModule = {
                 'Anonymize all data (contacts are always anonymized)',
               default: true,
             })
+            .option('subset', {
+              type: 'string',
+              description:
+                'Export subset: full for all data, demo for subset (400 contacts, 20 latest callouts)',
+              default: 'full',
+              choices: ['full', 'demo'],
+            })
             .option('outputDir', {
               type: 'string',
               description:
@@ -40,21 +47,23 @@ export const databaseCommand: CommandModule = {
           );
           return exportDatabase(argv.dryRun, argv.type as 'json' | 'sql', {
             anonymize: argv.anonymize,
-            subset: 'full',
+            subset: argv.subset as 'full' | 'demo',
             outputDir: argv.outputDir,
           });
         },
       })
       .command({
-        command: 'import <filePath>',
+        command: 'import [filePath]',
         describe:
-          'Import database from JSON or SQL dump file (auto-detects format from extension)',
+          'Import database from JSON or SQL dump file (defaults to test data dump)',
         builder: (yargs) =>
           yargs
             .positional('filePath', {
               type: 'string',
-              description: 'Path to the dump file (.json or .sql)',
-              demandOption: true,
+              description:
+                'Path to the dump file (.json or .sql). Defaults to test data dump if not provided',
+              default:
+                '../../packages/test-utils/database-dumps/database-dump.json',
             })
             .option('type', {
               type: 'string',
@@ -101,49 +110,6 @@ export const databaseCommand: CommandModule = {
             argv.calloutSlug as string,
             argv.filePath as string
           );
-        },
-      })
-      .command({
-        command: 'seed',
-        describe: 'Seed the database with test data from a JSON dump file',
-        builder: (yargs) =>
-          yargs
-            .option('dryRun', {
-              type: 'boolean',
-              description: 'Run without making changes',
-              default: false,
-            })
-            .option('filePath', {
-              type: 'string',
-              description: 'Full path to the JSON dump file',
-              default:
-                '../../packages/test-utils/database-dumps/database-dump.json',
-            }),
-        handler: async (argv) => {
-          const { seed } = await import('../actions/database/seed.js');
-          return seed(argv.dryRun, argv.filePath);
-        },
-      })
-      .command({
-        command: 'export-demo',
-        describe:
-          'Export demo database with subset of data (400 contacts, 20 latest callouts)',
-        builder: (yargs) =>
-          yargs
-            .option('dryRun', {
-              type: 'boolean',
-              description: 'Run without making changes',
-              default: false,
-            })
-            .option('type', {
-              type: 'string',
-              description: 'Export type: json or sql',
-              default: 'json',
-              choices: ['json', 'sql'],
-            }),
-        handler: async (argv) => {
-          const { exportDemo } = await import('../actions/database/demo.js');
-          return exportDemo(argv.dryRun, argv.type as 'json' | 'sql');
         },
       }),
   handler: () => {},
