+++ packages/core/src/tools/database/anonymisers/types.ts
@@ -1,0 +1,11 @@
+/**
+ * Types and interfaces for database dump operations
+ */
+
+/**
+ * Represents a database dump structure
+ * Maps table names to arrays of records
+ */
+export interface DatabaseDump {
+  [tableName: string]: any[];
+}
